requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
Django==5.0.2
django-appconf==1.0.6
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.1.0
easy-thumbnails==2.8.5
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
pilkit==3.0
pillow==10.2.0
pytz==2024.1
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
